+++ franka_example_controllers/include/franka_example_controllers/joint_position_example_controller.hpp
@@ -42,11 +42,12 @@ class JointPositionExampleController : public controller_interface::ControllerIn
 
  private:
   std::string arm_id_;
+  bool is_gazebo_{false};
+  std::string robot_description_;
   const int num_joints = 7;
   std::array<double, 7> initial_q_{0, 0, 0, 0, 0, 0, 0};
   const double trajectory_period{0.001};
   double elapsed_time_ = 0.0;
-  std::string arm_id{"panda"};
   const std::string k_HW_IF_INITIAL_POSITION = "initial_joint_position";
 
   bool initialization_flag_{true};