+++ franka_example_controllers/src/joint_position_example_controller.cpp
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include <franka_example_controllers/joint_position_example_controller.hpp>
+#include <franka_example_controllers/robot_utils.hpp>
 
 #include <cassert>
 #include <cmath>
@@ -37,8 +38,15 @@ controller_interface::InterfaceConfiguration
 JointPositionExampleController::state_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
-  for (int i = 1; i <= num_joints; ++i) {
-    config.names.push_back(arm_id_ + "_joint" + std::to_string(i) + "/" + k_HW_IF_INITIAL_POSITION);
+  if (!is_gazebo_) {
+    for (int i = 1; i <= num_joints; ++i) {
+      config.names.push_back(arm_id_ + "_joint" + std::to_string(i) + "/" +
+                             k_HW_IF_INITIAL_POSITION);
+    }
+  } else {
+    for (int i = 1; i <= num_joints; ++i) {
+      config.names.push_back(arm_id_ + "_joint" + std::to_string(i) + "/position");
+    }
   }
   return config;
 }
@@ -69,8 +77,8 @@ controller_interface::return_type JointPositionExampleController::update(
 
 CallbackReturn JointPositionExampleController::on_init() {
   try {
-    auto_declare<std::string>("arm_id", "panda");
-
+    auto_declare<bool>("gazebo", false);
+    auto_declare<std::string>("robot_description", "");
   } catch (const std::exception& e) {
     fprintf(stderr, "Exception thrown during init stage with message: %s \n", e.what());
     return CallbackReturn::ERROR;
@@ -80,7 +88,21 @@ CallbackReturn JointPositionExampleController::on_init() {
 
 CallbackReturn JointPositionExampleController::on_configure(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  arm_id_ = get_node()->get_parameter("arm_id").as_string();
+  is_gazebo_ = get_node()->get_parameter("gazebo").as_bool();
+
+  auto parameters_client =
+      std::make_shared<rclcpp::AsyncParametersClient>(get_node(), "/robot_state_publisher");
+  parameters_client->wait_for_service();
+
+  auto future = parameters_client->get_parameters({"robot_description"});
+  auto result = future.get();
+  if (!result.empty()) {
+    robot_description_ = result[0].value_to_string();
+  } else {
+    RCLCPP_ERROR(get_node()->get_logger(), "Failed to get robot_description parameter.");
+  }
+
+  arm_id_ = robot_utils::getRobotNameFromDescription(robot_description_, get_node()->get_logger());
 
   return CallbackReturn::SUCCESS;
 }