+++ CHANGELOG.md
@@ -1,11 +1,29 @@
 # Changelog
 
+## 0.1.15 - 2024-06-21
+
+Requires libfranka >= 0.13.2 and franka_description >= 0.3.0 requires ROS 2 Humble
+
+* franka\_gazebo\_bringup: Released and supports joint position, velocity and effort commands
+* franka\_ign\_ros2\_control: ROS 2 hardware interface for gazebo controller. Modified to add gravity torques for Franka robots.
+
+## 0.1.14 - 2024-05-13
+
+Requires libfranka >= 0.13.2, and franka_description >= 0.2.0 requires ROS 2 Humble
+
+* devcontainer: install pinocchio dependency from ros-humble-pinocchio apt package
+* feature: Added error recovery action to ROS 2 node
+* removed: hard-coded panda robot references
+* removed: franka_description package
+* using the franka_description standalone package https://github.com/frankaemika/franka_description
+* franka_hardware prefixes the robot_state and robot model state interfaces with the read robot name from the urdf. 
+
 ## 0.1.13 - 2024-01-18
 
 Requires libfranka >= 0.13.2, requires ROS 2 Humble
 
-* update libfranka dependency on devcontainer to 0.13.3(requires system image 5.5.0)
-* fix .devcontainer typo
+* update libfranka dependency in devcontainer to 0.13.3(requires system image 5.5.0)
+* fix devcontainer typo
 
 ## 0.1.12 - 2024-01-12